requirements.txt
streamlit
openai-whisper
sounddevice
scipy
